requirements-test.txt
django>=1.9.1
coverage
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
twilio==5.4.0
nose==1.3.7
djangorestframework==3.3.3

# Additional test requirements go here
